% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_weights.R
\name{assign_weights}
\alias{assign_weights}
\title{Assign weights (i.e., likelihoods) to each of the value}
\usage{
assign_weights(
  var,
  t,
  data,
  data_type = c("infection", "symptom onset", "confirmation")
)
}
\arguments{
\item{var}{A vector of state variables from simulation}

\item{t}{current time}

\item{data}{data on the times of the infection transmission process (e.g., infection, symptom onset, or confirmation)}

\item{data_type}{c("infection", "symptom onset", "confirmation")}
}
\description{
The \code{assign_weights()} calculate the likelihoods assuming case incidences are Poisson-distributed random numbers
}
\examples{
 wt <- assign_weights(var = latent_var, t = t, data = data, data_type = type)
}
