% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trace.R
\name{extract_trace}
\alias{extract_trace}
\title{Extract results from particle filtering and backward sampling (trace)}
\usage{
extract_trace(
  params = NULL,
  y = NULL,
  beta0 = NULL,
  data = NULL,
  data_type = c("infection", "symptom onset", "confirmation"),
  rep = 1,
  npart = 100,
  tend = 200,
  dt = 0.2,
  error_pdf = c("pois", "negbin", "binom"),
  negbin_size = 5,
  binom_prob = 0.8,
  systematic_resampling = FALSE
)
}
\arguments{
\item{y}{lstate variables}

\item{data}{data to calculate likelihood against}

\item{data_type}{infection, symptom onset, or confirmation}

\item{theta}{Parameter values}
}
\description{
The \code{extract_trace()} extracts trace
}
\examples{
res <- extract_trace()
}
