% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{gen}
\alias{gen}
\alias{gen.pframe_1d}
\alias{gen.default}
\title{Generate Method}
\usage{
gen(x, ...)

\method{gen}{pframe_1d}(...)

\method{gen}{default}(...)
}
\arguments{
\item{x}{is object, i.e. pframe_1d}
}
\value{
Description of the object you will be returning.
\itemize{
 \item{"Attribute 1" :}{Name}
 \item{"Attribute 2" :}{Call}
 \item{"List Item 1" :}{  Vector}
}
}
\description{
Generate random values from empty vector with known dimension
}
\details{
If you like to write a lot, you can always describe more here.
}
\note{
You can also leave an important note here
}
\examples{
a+b
}
\references{
Links can be useful \url{http://en.wikipedia.org/wiki/Fermat's_little_theorem}
}
\seealso{
\code{\link{particle}} This part tells you to refer to other code
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
