% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_x0.R
\name{init_x0}
\alias{init_x0}
\alias{init_x0.pframe_1d}
\alias{init_x0.default}
\title{Initialise prior distributions for pframe_1d}
\usage{
init_x0(obj, ...)

\method{init_x0}{pframe_1d}(obj, dist_x0 = "gaussian", params_x0 = list(mean
  = 0, sd = 1), ...)

\method{init_x0}{default}(obj)
}
\arguments{
\item{obj}{Object of pframe_1d class}

\item{...}{optional arguments}

\item{dist_x0}{Prior distribution (gaussian,t, gamma, etc)}

\item{obj}{Object of pframe_1d}
}
\value{
Updates or initialise pframe_1d with prior distribution.
}
\description{
\code{init_x0} is a method
to initialise the prior state of particles in \code{x0} in pframe_1d object.
}
\note{
\code{init_x0} is not needed if all arguments are provided by the constructor pframe_1d.
This function is a convenience for user. A proficient user will be able to update or initialise variables by hand.
}
\examples{
init_x0(pframe_1d())

}
\seealso{
\code{\link{project}}, \code{\link{update}} and \code{\link{resample}}
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
