% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlik.R
\name{maxlik}
\alias{maxlik}
\title{Estimate parameter values using the maximum likelihood approach}
\usage{
maxlik(
  theta = theta0,
  y = y0,
  data = Rt_data,
  data_type = c("infection", "symptom onset", "confirmation"),
  tend = 200,
  dt = 0.2
)
}
\arguments{
\item{theta}{Parameters}

\item{y}{A vector of state variables (e.g., c(S,E1,E2,I,X))}

\item{data}{data to calculate the likelihoods against}

\item{data_type}{data to calculate the likelihoods against}
}
\description{
The \code{maxlik()} estimates the parameter values via the maximum likelihood approach
}
\examples{
sample <- maxlik(theta = theta, y = y0, data = Rt_data, data_type = "infection", tend = 200, dt = 0.2)
}
