% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglik.R
\name{negloglik}
\alias{negloglik}
\title{Assign weights (i.e., likelihoods) to each of the value}
\usage{
negloglik(
  pars = NULL,
  theta = NULL,
  y = NULL,
  tbegin = NULL,
  tend = NULL,
  dt = NULL,
  data = NULL,
  data_type = c("infection", "symptom onset", "confirmation")
)
}
\arguments{
\item{data}{data on the times of the infection transmission process (e.g., infection, symptom onset, or confirmation)}

\item{data_type}{c("infection", "symptom onset", "confirmation")}

\item{var}{A vector of state variables from simulation}

\item{t}{current time}
}
\description{
The \code{negloglik()} calculate the negative log likelihoods assuming case incidences are Poisson-distributed random numbers
}
\examples{
 nll <- negloglik(var = latent_var, t = t, data = data, data_type = type)
}
