% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle.R
\name{particle}
\alias{particle}
\alias{particle.pframe_1d}
\alias{particle.default}
\title{Running Particle Filter}
\usage{
particle(obj, ...)

\method{particle}{pframe_1d}(obj, ...)

\method{particle}{default}(obj)
}
\arguments{
\item{obj}{Object of pframe_1d class}
}
\value{
Object of pframe_1d class
}
\description{
Wrapper for project, update, resample to run through whole entire particle filter,
given the full \code{y} is know.
}
\examples{
project(init_x0(pframe_1d()))

}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
