% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter-package.r
\docType{package}
\name{pfilter}
\alias{pfilter}
\alias{pfilter-package}
\title{pfilter.}
\description{
Lightweight particle filtering algorithms using S3 classes
}
\section{Wow}{
NA
}

\section{Wow2}{
NA
}

\section{Wow3}{
NA
}

