% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pframe_1d.R
\name{pframe_1d}
\alias{pframe_1d}
\title{Constructor for pframe_1d class}
\usage{
pframe_1d(N = 100, y = c(1, 1, 1), resample = TRUE, smooth = TRUE,
  sequential = TRUE, dist_f = "gaussian", params_f = list(),
  dist_g = "gaussian", params_g = list(sd = 1),
  resamp_type = "multinomial", ...)
}
\arguments{
\item{N}{Number of particles}

\item{y}{Full data vector}

\item{resample}{Boolean for resampling}

\item{smooth}{Boolean for smooth}

\item{dist_f}{Distribution for project method (gaussian,t, gamma, etc)}

\item{params_f}{list of parameters for project method}

\item{dist_g}{Likelihood for update method (gaussian,t, gamma, etc)}

\item{params_g}{list of parameters for update method}

\item{tau}{Number of time points}
}
\value{
An object instantiated from pframe_1d class. Methods include init ,project, update, resample and
metric.
}
\description{
\code{pframe_1d} is a constructor function that generates an instance of the S3 pframe_1d class.
By passing in a set of parameters (or using defaults), a pframe_1d object is generated
where filtering methods can be applied on. pframe_1d is a frameholder that creates empty variables, vectors and matrices;
some of which are expected, whereas others are optional.
}
\note{
\code{y} is the full data vector, therefore, we assume that filtering is not essential. Therefore,
it is preferred to run smoothing if \code{y} is available.
}
\examples{
obj1<-pframe_1d()
}
\seealso{
\code{\link{project}}, \code{\link{update}} and \code{\link{resample}}
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
