% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pframe_1d}
\alias{plot.pframe_1d}
\title{Plot object with class pframe_1d}
\usage{
\method{plot}{pframe_1d}(obj, clean = FALSE, ...)
}
\arguments{
\item{obj}{Object of pframe_1d class}

\item{clean}{Print a prettified version of plot}
}
\value{
Graphical plot of obj
}
\description{
Plot method for object with class pframe_1d. The method prints the trend of each
particle trajectory.
}
\examples{
plot(particle(pframe_1d()))
set.seed(1)
x<-c(1,2,3,4,4,4,5,4,5,4,4,4,4,4,4) #true means of state model
plot(particle(pframe_1d(N=100,smooth=FALSE,y=rnorm(length(x),mean=x))))

}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
