% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_var.R
\name{plot_latent_var}
\alias{plot_latent_var}
\title{This function plots latent variables}
\usage{
plot_latent_var(
  sim_res = pf_result,
  data = Rt_data,
  data_type = c("infection", "symptom onset", "confirmation")
)
}
\arguments{
\item{sim_res}{Simulation results}

\item{data}{data to compare}

\item{data_type}{data based on the date of infection, symptom onset, or confirmation}
}
\description{
The \code{plot_latent_var()} estimates posterior distribution using the particle filtering method
population for both sexes and incidence rate
}
\examples{
pf <- particle_filter(data_type = "infection", npart = 1e3, dt = 0.1); plot_latent_var(sim_res = pf)
}
