% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_var_pl.R
\name{plot_latent_var_pl}
\alias{plot_latent_var_pl}
\title{This function plots latent variables}
\usage{
plot_latent_var_pl(
  sim_res = pf_result,
  data = Rt_data,
  data_type = c("infection", "symptom onset", "confirmation")
)
}
\arguments{
\item{sim_res}{Simulation results}

\item{data}{data to compare}

\item{data_type}{data based on the date of infection, symptom onset, or confirmation}
}
\description{
The \code{plot_latent_var_pl()} estimates posterior distribution using the particle filtering method
population for both sexes and incidence rate
}
\examples{
plot_latent_var_pl(sim_res = pf)
}
