% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\alias{project.pframe_1d}
\alias{project.default}
\title{Project particles}
\usage{
project(obj, ...)

\method{project}{pframe_1d}(obj, dist_f, params_f, ...)

\method{project}{default}(obj)
}
\arguments{
\item{obj}{Object of pframe_1d class}
}
\value{
Object of pframe_1d class
}
\description{
Method to project particles forward using state model f(xn+1|xn). Given a set of particles and
state model, predicts where particles may be.
}
\details{
In genetic algorithms, these functions are isomorphic to mutation/ prediction
}
\examples{
project(init_x0(pframe_1d()))
}
\seealso{
\code{\link{particle}}
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
