% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.R
\name{resamp}
\alias{resamp}
\alias{resamp.pframe_1d}
\alias{resamp.default}
\title{Resample particles}
\usage{
resamp(obj, ...)

\method{resamp}{pframe_1d}(obj, ...)

\method{resamp}{default}(obj)
}
\arguments{
\item{obj}{Object of pframe_1d class}
}
\value{
Object of pframe_1d class
}
\description{
Resample particles based on computed weights.
}
\details{
In genetic algorithms, these functions are part of selection/updating
}
\note{
Since resampling is computationally costly, it be chosen to be done only "sometimes".
This is why the effective sample size is used, \code{N_eff}
}
\examples{
resamp(update(project(init_x0(pframe_1d()))))
}
\seealso{
\code{\link{particle}},\code{\link{update}}
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
