% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run particle filtering models and sample trajectories of variables}
\usage{
run_model(
  params = theta,
  y = y0,
  data = Rt_data,
  data_type = c("infection", "symptom onset", "confirmation"),
  rep = 1,
  npart = 100,
  tend = 200,
  dt = 0.2,
  systematic_resampling = FALSE
)
}
\arguments{
\item{y}{lstate variables}

\item{data}{data to calculate likelihood against}

\item{data_type}{infection, symptom onset, or confirmation}

\item{theta}{Parameter values}
}
\description{
The \code{run_model()} runs particle filtering models
}
\examples{
res <- run_model()
}
