% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_step.R
\name{run_step}
\alias{run_step}
\title{Runs a step (day in this case)}
\usage{
run_step(y, times, params)
}
\arguments{
\item{y}{A vector of state variables}

\item{times}{Times over which ODE is integrated (1 day in our case)}

\item{params}{A vector of parameter values (i.e., sigma, epsilon, gamma, R)}
}
\description{
The \code{step_run} returns the rate of change of the SE1E2IR model that can be integrated using deSolve package
}
\examples{
res <- sepir(t, y, param)
}
