% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepir.R
\name{sepir}
\alias{sepir}
\title{Implements the rate of change of the SE1E2IR model for integration using deSolve}
\usage{
sepir(t, y, param)
}
\arguments{
\item{t}{A vector of times for integration}

\item{y}{A vector of state variables}

\item{param}{A vector of parameter values (i.e., sigma, epsilon, gamma, R)}
}
\description{
The \code{sepir} returns the rate of change of the SE1E2IR model that can be integrated using deSolve package
}
\examples{
res <- sepir(t, y, param)
}
