% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\alias{show.pframe_1d}
\alias{show.default}
\title{Show method}
\usage{
show(obj, ...)

\method{show}{pframe_1d}(obj, dist_f = rnorm, ...)

\method{show}{default}(obj)
}
\arguments{
\item{x}{is object, i.e. pframe_1d}
}
\value{
Description of the object you will be returning.
}
\description{
Show object of pframe_1d class
}
\details{
In genetic algorithms, these functions are isomorphic to mutation/ prediction
}
\note{
You can also leave an important note here
}
\examples{
a+b
}
\references{
Links can be useful \url{http://en.wikipedia.org/wiki/Fermat's_little_theorem}
}
\seealso{
\code{\link{particle}} This part tells you to refer to other code
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
