% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_resampling.R
\name{systematic_resampling}
\alias{systematic_resampling}
\title{Re-sample systematically (by comparing to the cumulative distribution)}
\usage{
systematic_resampling(weights)
}
\arguments{
\item{weights}{A vector of weights}
}
\description{
The \code{systematic_resampling()} a
population for both sexes and incidence rate
}
\examples{
id <- systematic_resampling
}
