% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update.pframe_1d}
\alias{update.default}
\title{Update weights}
\usage{
update(obj, ...)

\method{update}{pframe_1d}(obj, dist_g, params_g, ...)

\method{update}{default}(obj)
}
\arguments{
\item{obj}{Object of pframe_1d class}
}
\value{
Object of pframe_1d class
}
\description{
Each particle is given a different weight based on how close each particle is to
the data point by using the likelihood.
}
\details{
In genetic algorithms, these functions are part of selection/updating
}
\note{
One must becareful of particle degeneracy. Occasionally, all weight is given to one particle
only. This usually occurs when the state model does not conform with the data.
}
\examples{
update(project(init_x0(pframe_1d())))
}
\seealso{
\code{\link{particle}}
}
\author{
Justin Thong \email{justinthong93@gmail.com}
}
